requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
click-log==0.3.2
idna==2.9
pkg-resources==0.0.0
pysonofflan==0.3.0
requests==2.23.0
six==1.14.0
sonoff-python==0.2.1
urllib3==1.25.9
websocket-client==0.57.0
websockets==8.1
